move ConanHostedHelper back out
