+++ nexus-repository-conan/src/main/java/org/sonatype/repository/conan/internal/hosted/ConanHostedHelper.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.repository.conan.internal.orient.hosted;
+package org.sonatype.repository.conan.internal.hosted;
 
 import org.sonatype.nexus.repository.view.matchers.token.TokenMatcher;
 import org.sonatype.repository.conan.internal.AssetKind;
